refactor: ResultView 의 출력 책임 위임 및 Racing 결합 제거
