--- src/main/java/racing/ResultView.java
@@ -1,12 +1,0 @@
-package racing;
-
-public class ResultView {
-
-    public void printRacingState(Racing racing) {
-        racing.getCarList()
-              .forEach(car -> System.out.println(car.getTrack()));
-
-        System.out.println();
-    }
-
-}
