+++ src/main/java/racing/car/Car.java
@@ -1,0 +1,15 @@
+package racing.car;
+
+public class Car {
+
+    private int location = 0;
+
+    public int getLocation() {
+        return location;
+    }
+
+    public void move() {
+        location++;
+    }
+
+}
